requirements.txt
meshtastic
paho-mqtt
pypubsub
python-dotenv
pyserial